Disable haptic feedback on power menu items
Power menu items without long press actions should not have haptic feedback.
It not only makes zero sense but gives the user the impression that something
is not working correctly. Haptic feedback is generally used to notify the user
that an action has been triggered. If there's no action to trigger than there
should not be haptic feedback.

Change-Id: I60b4a2b58e75c1e0bfca0d271b4cdd4c83e5633d